所有页面展示分类

+++ src/main/java/com/bookstore/bookstore/service/info/ClassIficationInfo.java
@@ -7,7 +7,7 @@ import java.util.Collection;
 @Data
 public class ClassIficationInfo {
 
-    private int classID;
+    private Long classID;
 
     private  String className;
